Cleaned up code for linear PC DMFT and organised files
